% Auto-generated: do not edit by hand
\name{bookmarks}

\alias{bookmarks}

\title{Bookmarks component}

\description{

}

\usage{
bookmarks(id=NULL, editingEnabled=NULL, label=NULL, position=NULL,
visible=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{editingEnabled}{editingEnabled (boolean; optional): Indicates whether the bookmarks are able to be edited.}

\item{label}{The widget's default label.}

\item{position}{position of the basemap gallery. Defaults to top-right.
Possible values: 'bottom-leading', 'bottom-left', 'bottom-right', 'bottom-trailing', 'top-leading', 'top-left', 'top-right', 'top-trailing'. Default: 'top-right'.}

\item{visible}{Indicates whether the widget is visible.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

