% Auto-generated: do not edit by hand
\name{editor}

\alias{editor}

\title{Editor component}

\description{

}

\usage{
editor(id=NULL, label=NULL, position=NULL, visible=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{label}{The widget's default label.}

\item{position}{position of the basemap gallery. Defaults to top-right.}

\item{visible}{Indicates whether the widget is visible.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

