% Auto-generated: do not edit by hand
\name{geoJSONLayer}

\alias{geoJSONLayer}

\title{GeoJSONLayer component}

\description{

}

\usage{
geoJSONLayer(id=NULL, url=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{url}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

