% Auto-generated: do not edit by hand
\name{mapViewBase}

\alias{mapViewBase}

\title{MapViewBase component}

\description{

}

\usage{
mapViewBase(id=NULL, animation=NULL, basemap=NULL, center=NULL,
style=NULL, type=NULL, viewPoint=NULL, widthBreakpoint=NULL,
zoom=NULL)
}

\arguments{
\item{id}{}

\item{animation}{The ID used to identify this component in Dash callbacks.}

\item{basemap}{}

\item{center}{}

\item{style}{}

\item{type}{}

\item{viewPoint}{}

\item{widthBreakpoint}{}

\item{zoom}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

