% Auto-generated: do not edit by hand
\name{webMap}

\alias{webMap}

\title{WebMap component}

\description{

}

\usage{
webMap(children=NULL, id=NULL, basemap=NULL, breakpoints=NULL,
center=NULL, constraints=NULL, extent=NULL,
heightBreakpoint=NULL, portalItem=NULL, style=NULL,
widthBreakpoint=NULL, zoom=NULL)
}

\arguments{
\item{children}{}

\item{id}{}

\item{basemap}{}

\item{breakpoints}{}

\item{center}{}

\item{constraints}{}

\item{extent}{}

\item{heightBreakpoint}{}

\item{portalItem}{}

\item{style}{}

\item{widthBreakpoint}{}

\item{zoom}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

