% Auto-generated: do not edit by hand
\name{baseMapGallery}

\alias{baseMapGallery}

\title{BaseMapGallery component}

\description{

}

\usage{
baseMapGallery(id=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

