% Auto-generated: do not edit by hand
\name{mapView}

\alias{mapView}

\title{MapView component}

\description{

}

\usage{
mapView(id=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

