% Auto-generated: do not edit by hand
\name{mapView}

\alias{mapView}

\title{MapView component}

\description{

}

\usage{
mapView(id=NULL, animation=NULL, basemap=NULL, center=NULL,
style=NULL, type=NULL, viewPoint=NULL, widthBreakpoint=NULL,
zoom=NULL)
}

\arguments{
\item{id}{Character. }

\item{animation}{Named list. The ID used to identify this component in Dash callbacks.}

\item{basemap}{Character. }

\item{center}{Unnamed list. }

\item{style}{Named list. }

\item{type}{Named list. }

\item{viewPoint}{Named list. }

\item{widthBreakpoint}{Character. }

\item{zoom}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

